Got rid of some tests
