+++ cypress/e2e/yves/order-management/cart-custom-order-reference.cy.ts
@@ -4,7 +4,6 @@ import { CartPage, MultiCartPage, CustomOrderReferenceCartPage } from '@pages/yv
 import { CustomerLoginScenario } from '@scenarios/yves';
 
 describe('cart item quantity', { tags: ['@order-management'] }, (): void => {
-  const cartPage = container.get(CartPage);
   const multiCartPage = container.get(MultiCartPage);
   const customOrderReferenceCartPage = container.get(CustomOrderReferenceCartPage);
   const customerLoginScenario = container.get(CustomerLoginScenario);
@@ -26,43 +25,9 @@ describe('cart item quantity', { tags: ['@order-management'] }, (): void => {
   it('customer should be able to add a custom order reference', (): void => {
     multiCartPage.visit();
     multiCartPage.selectCart({ name: dynamicFixtures.quote.name });
-    cartPage.visitCartWithItems();
 
     customOrderReferenceCartPage.addCustomOrderReferenceInput(staticFixtures.reference);
 
     customOrderReferenceCartPage.getCustomOrderReferenceInput().should('have.value', staticFixtures.reference);
-  });
-
-  it('adding a custom order reference to cart should not scroll the page to the top', (): void => {
-    cy.viewport(800, 800);
-
-    multiCartPage.visit();
-    multiCartPage.selectCart({ name: dynamicFixtures.quote.name });
-    cartPage.visitCartWithItems();
-
-    customOrderReferenceCartPage.addCustomOrderReferenceInput(staticFixtures.reference);
-
-    cy.window().then((win) => {
-      expect(win.scrollY).not.equal(0);
-    });
-  });
-
-  it('adding custom order reference should not reload the whole cart block', (): void => {
-    multiCartPage.visit();
-    multiCartPage.selectCart({ name: dynamicFixtures.quote.name });
-    cartPage.visitCartWithItems();
-
-    customOrderReferenceCartPage.addCustomOrderReferenceInput(staticFixtures.reference);
-
-    cartPage
-      .getCheckoutButton()
-      .invoke('append', '<div class="added-html">Added HTML</div>')
-      .then(() => {
-        customOrderReferenceCartPage.addCustomOrderReferenceInput(staticFixtures.reference);
-
-        cartPage.getCheckoutButton().then(($button) => {
-          expect($button.find('.added-html')).to.exist;
-        });
-      });
   });
 });
