--- cypress/fixtures/suite/yves/cart/dynamic-cart-item-quantity.json
@@ -1,110 +1,0 @@
-{
-  "data": {
-    "type": "dynamic-fixtures",
-    "attributes": {
-      "synchronize": true,
-      "operations": [
-        {
-          "type": "transfer",
-          "name": "LocaleTransfer",
-          "key": "locale",
-          "arguments": { "id_locale": 66, "locale_name": "en_US" }
-        },
-        {
-          "type": "helper",
-          "name": "haveCustomer",
-          "key": "customer",
-          "arguments": [{ "locale": "#locale", "password": "change123" }]
-        },
-        {
-          "type": "helper",
-          "name": "confirmCustomer",
-          "key": "confirmedCustomer",
-          "arguments": ["#customer"]
-        },
-        {
-          "type": "transfer",
-          "name": "StoreTransfer",
-          "key": "store",
-          "arguments": { "id_store": 1 }
-        },
-        {
-          "type": "helper",
-          "name": "haveFullProduct",
-          "key": "product"
-        },
-        {
-          "type": "helper",
-          "name": "havePriceProduct",
-          "arguments": [
-            {
-              "skuProductAbstract": "#product.abstract_sku",
-              "skuProduct": "#product.sku",
-              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "arguments": ["#store", { "sku": "#product.sku", "isNeverOutOfStock": "1" }]
-        },
-        {
-          "type": "transfer",
-          "name": "ProductImageTransfer",
-          "key": "productImage",
-          "arguments": {
-            "externalUrlSmall": "https://images.icecat.biz/img/gallery_mediums/30691822_1486.jpg",
-            "externalUrlLarge": "https://images.icecat.biz/img/gallery/30691822_1486.jpg"
-          }
-        },
-        {
-          "type": "helper",
-          "name": "haveProductImageSet",
-          "arguments": [
-            {
-              "name": "default",
-              "idProduct": "#product.id_product_concrete",
-              "idProductAbstract": "#product.fk_product_abstract",
-              "productImages": ["#productImage"]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePersistentQuote",
-          "key": "quote1",
-          "arguments": [
-            {
-              "customer": "#customer",
-              "items": [
-                {
-                  "sku": "#product.sku",
-                  "abstractSku": "#product.abstract_sku",
-                  "quantity": 2
-                }
-              ]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePersistentQuote",
-          "key": "quote2",
-          "arguments": [
-            {
-              "customer": "#customer",
-              "items": [
-                {
-                  "sku": "#product.sku",
-                  "abstractSku": "#product.abstract_sku",
-                  "quantity": 2
-                }
-              ]
-            }
-          ]
-        }
-      ]
-    }
-  }
-}
