--- cypress/fixtures/suite/yves/cart/dynamic-cart-item-remove.json
@@ -1,191 +1,0 @@
-{
-  "data": {
-    "type": "dynamic-fixtures",
-    "attributes": {
-      "synchronize": true,
-      "operations": [
-        {
-          "type": "transfer",
-          "name": "LocaleTransfer",
-          "key": "locale",
-          "arguments": { "id_locale": 66, "locale_name": "en_US" }
-        },
-        {
-          "type": "helper",
-          "name": "haveCustomer",
-          "key": "customer",
-          "arguments": [{ "locale": "#locale", "password": "change123" }]
-        },
-        {
-          "type": "helper",
-          "name": "confirmCustomer",
-          "key": "confirmedCustomer",
-          "arguments": ["#customer"]
-        },
-        {
-          "type": "transfer",
-          "name": "StoreTransfer",
-          "key": "store",
-          "arguments": { "id_store": 1 }
-        },
-        {
-          "type": "helper",
-          "name": "haveFullProduct",
-          "key": "product1"
-        },
-        {
-          "type": "helper",
-          "name": "haveFullProduct",
-          "key": "product2"
-        },
-        {
-          "type": "helper",
-          "name": "haveFullProduct",
-          "key": "product3"
-        },
-        {
-          "type": "helper",
-          "name": "havePriceProduct",
-          "arguments": [
-            {
-              "skuProductAbstract": "#product1.abstract_sku",
-              "skuProduct": "#product1.sku",
-              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePriceProduct",
-          "arguments": [
-            {
-              "skuProductAbstract": "#product2.abstract_sku",
-              "skuProduct": "#product2.sku",
-              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePriceProduct",
-          "arguments": [
-            {
-              "skuProductAbstract": "#product3.abstract_sku",
-              "skuProduct": "#product3.sku",
-              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "arguments": ["#store", { "sku": "#product1.sku", "isNeverOutOfStock": "1" }]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "arguments": ["#store", { "sku": "#product2.sku", "isNeverOutOfStock": "1" }]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "arguments": ["#store", { "sku": "#product3.sku", "isNeverOutOfStock": "1" }]
-        },
-        {
-          "type": "transfer",
-          "name": "ProductImageTransfer",
-          "key": "productImage",
-          "arguments": {
-            "externalUrlSmall": "https://images.icecat.biz/img/gallery_mediums/30691822_1486.jpg",
-            "externalUrlLarge": "https://images.icecat.biz/img/gallery/30691822_1486.jpg"
-          }
-        },
-        {
-          "type": "helper",
-          "name": "haveProductImageSet",
-          "arguments": [
-            {
-              "name": "default",
-              "idProduct": "#product1.id_product_concrete",
-              "idProductAbstract": "#product1.fk_product_abstract",
-              "productImages": ["#productImage"]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductImageSet",
-          "arguments": [
-            {
-              "name": "default",
-              "idProduct": "#product2.id_product_concrete",
-              "idProductAbstract": "#product2.fk_product_abstract",
-              "productImages": ["#productImage"]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductImageSet",
-          "arguments": [
-            {
-              "name": "default",
-              "idProduct": "#product3.id_product_concrete",
-              "idProductAbstract": "#product3.fk_product_abstract",
-              "productImages": ["#productImage"]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePersistentQuote",
-          "key": "quote1",
-          "arguments": [
-            {
-              "customer": "#customer",
-              "items": [
-                {
-                  "sku": "#product1.sku",
-                  "abstractSku": "#product1.abstract_sku",
-                  "quantity": 1
-                },
-                {
-                  "sku": "#product2.sku",
-                  "abstractSku": "#product2.abstract_sku",
-                  "quantity": 1
-                },
-                {
-                  "sku": "#product3.sku",
-                  "abstractSku": "#product3.abstract_sku",
-                  "quantity": 1
-                }
-              ]
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "havePersistentQuote",
-          "key": "quote2",
-          "arguments": [
-            {
-              "customer": "#customer",
-              "items": [
-                {
-                  "sku": "#product1.sku",
-                  "abstractSku": "#product1.abstract_sku",
-                  "quantity": 1
-                },
-                {
-                  "sku": "#product2.sku",
-                  "abstractSku": "#product2.abstract_sku",
-                  "quantity": 2
-                }
-              ]
-            }
-          ]
-        }
-      ]
-    }
-  }
-}
