+++ cypress/fixtures/suite/yves/cart/dynamic-cart-item.json
@@ -6,20 +6,9 @@
       "operations": [
         {
           "type": "transfer",
-          "name": "StoreTransfer",
-          "key": "store",
-          "arguments": { "id_store": 1 }
-        },
-        {
-          "type": "transfer",
           "name": "LocaleTransfer",
           "key": "locale",
           "arguments": { "id_locale": 66, "locale_name": "en_US" }
-        },
-        {
-          "type": "helper",
-          "name": "haveCountry",
-          "key": "country"
         },
         {
           "type": "helper",
@@ -29,20 +18,57 @@
         },
         {
           "type": "helper",
-          "name": "haveCustomerAddress",
-          "key": "address",
-          "arguments": [{ "fkCustomer": "#customer.id_customer", "fkCountry": "#country.id_country" }]
-        },
-        {
-          "type": "helper",
           "name": "confirmCustomer",
           "key": "confirmedCustomer",
           "arguments": ["#customer"]
         },
         {
+          "type": "transfer",
+          "name": "StoreTransfer",
+          "key": "store",
+          "arguments": { "id_store": 1 }
+        },
+        {
           "type": "helper",
           "name": "haveFullProduct",
           "key": "product1"
+        },
+        {
+          "type": "helper",
+          "name": "haveFullProduct",
+          "key": "product2"
+        },
+        {
+          "type": "helper",
+          "name": "havePriceProduct",
+          "arguments": [
+            {
+              "skuProductAbstract": "#product1.abstract_sku",
+              "skuProduct": "#product1.sku",
+              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "havePriceProduct",
+          "arguments": [
+            {
+              "skuProductAbstract": "#product2.abstract_sku",
+              "skuProduct": "#product2.sku",
+              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveProductInStockForStore",
+          "arguments": ["#store", { "sku": "#product1.sku", "isNeverOutOfStock": "1" }]
+        },
+        {
+          "type": "helper",
+          "name": "haveProductInStockForStore",
+          "arguments": ["#store", { "sku": "#product2.sku", "isNeverOutOfStock": "1" }]
         },
         {
           "type": "transfer",
@@ -67,11 +93,6 @@
         },
         {
           "type": "helper",
-          "name": "haveFullProduct",
-          "key": "product2"
-        },
-        {
-          "type": "helper",
           "name": "haveProductImageSet",
           "arguments": [
             {
@@ -84,42 +105,25 @@
         },
         {
           "type": "helper",
-          "name": "havePriceProduct",
-          "key": "product1WithPrice",
+          "name": "havePersistentQuote",
+          "key": "quote",
           "arguments": [
             {
-              "skuProductAbstract": "#product1.abstract_sku",
-              "skuProduct": "#product1.sku",
-              "moneyValue": {
-                "netAmount": 30000,
-                "grossAmount": 30000
-              }
+              "customer": "#customer",
+              "items": [
+                {
+                  "sku": "#product1.sku",
+                  "abstractSku": "#product1.abstract_sku",
+                  "quantity": 2
+                },
+                {
+                  "sku": "#product2.sku",
+                  "abstractSku": "#product2.abstract_sku",
+                  "quantity": 1
+                }
+              ]
             }
           ]
-        },
-        {
-          "type": "helper",
-          "name": "havePriceProduct",
-          "key": "product2WithPrice",
-          "arguments": [
-            {
-              "skuProductAbstract": "#product2.abstract_sku",
-              "skuProduct": "#product2.sku",
-              "moneyValue": { "netAmount": 30000, "grossAmount": 30000 }
-            }
-          ]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "key": "product1WithPriceAndStore",
-          "arguments": ["#store", { "sku": "#product1.sku", "isNeverOutOfStock": "1" }]
-        },
-        {
-          "type": "helper",
-          "name": "haveProductInStockForStore",
-          "key": "product2WithPriceAndStore",
-          "arguments": ["#store", { "sku": "#product2.sku", "isNeverOutOfStock": "1" }]
         }
       ]
     }
