--- cypress/fixtures/suite/yves/cart/static-cart-item-quantity.json
@@ -1,3 +1,0 @@
-{
-  "defaultPassword": "change123"
-}
