+++ cypress/fixtures/suite/yves/cart/static-cart-item.json
