+++ cypress/support/pages/yves/comment/cart/comment-cart-page.ts
@@ -21,8 +21,6 @@ export class CommentCartPage extends YvesPage {
     this.repository.getAddCommentForm().last().find('textarea').clear();
     this.repository.getAddCommentForm().last().find('textarea').type(params.message);
     this.repository.getAddCommentForm().last().find(this.repository.getAddCommentButtonSelector()).click();
-
-    this.repository.getCommentThreadAjaxLoader().should('be.not.visible');
   };
 
   update = (params: UpdateParams): void => {
@@ -30,8 +28,6 @@ export class CommentCartPage extends YvesPage {
 
     textarea.clear().type(params.newMessage);
     textarea.parent().find(this.repository.getUpdateCommentButtonSelector()).click();
-
-    this.repository.getCommentThreadAjaxLoader().should('be.not.visible');
   };
 
   remove = (params: RemoveParams): void => {
@@ -40,8 +36,6 @@ export class CommentCartPage extends YvesPage {
       .parent()
       .find(this.repository.getRemoveCommentButtonSelector())
       .click();
-
-    this.repository.getCommentThreadAjaxLoader().should('be.not.visible');
   };
 
   updateFirst = (params: UpdateFirstParams): void => {
@@ -49,14 +43,10 @@ export class CommentCartPage extends YvesPage {
     textarea.clear().type(params.message);
 
     this.getCommentThreadListSection().first().find(this.repository.getUpdateCommentButtonSelector()).click();
-
-    this.repository.getCommentThreadAjaxLoader().should('be.not.visible');
   };
 
   removeFirst = (): void => {
     this.getCommentThreadListSection().find(this.repository.getRemoveCommentButtonSelector()).click();
-
-    this.repository.getCommentThreadAjaxLoader().should('be.not.visible');
   };
 
   getCommentThreadListSection = (): Cypress.Chainable => {
