+++ cypress/support/pages/yves/comment/cart/repositories/suite-comment-cart-repository.ts
@@ -17,6 +17,4 @@ export class SuiteCommentCartRepository implements CommentCartRepository {
   }
   getCartUpsellingAjaxLoader = (): Cypress.Chainable =>
     cy.get('[data-qa="component cart-upselling"] [data-qa="component ajax-loader"]', { timeout: 10000 });
-  getCommentThreadAjaxLoader = (): Cypress.Chainable =>
-    cy.get('[data-qa="component ajax-loader comment-thread-ajax-loader"]', { timeout: 10000 });
 }
