+++ cypress/support/pages/yves/multi-cart/multi-cart-repository.ts
@@ -2,4 +2,5 @@ export interface MultiCartRepository {
   getCreateCartNameInput(): Cypress.Chainable;
   getCreateCartForm(): Cypress.Chainable;
   getQuoteTable(): Cypress.Chainable;
+  getCartUpsellingAjaxLoader(): Cypress.Chainable;
 }
