+++ cypress/support/pages/yves/multi-cart/repositories/b2b-mp-multi-cart-repository.ts
@@ -7,4 +7,6 @@ export class B2bMpMultiCartRepository implements MultiCartRepository {
   getCreateCartNameInput = (): Cypress.Chainable => cy.get('#quoteForm_name');
   getCreateCartForm = (): Cypress.Chainable => cy.get('form[name=quoteForm]');
   getQuoteTable = (): Cypress.Chainable => cy.get('[data-qa="component quote-table"]');
+  getCartUpsellingAjaxLoader = (): Cypress.Chainable =>
+    cy.get('[data-qa="component cart-upselling"] [data-qa="component ajax-loader"]', { timeout: 10000 });
 }
