+++ cypress/support/pages/yves/order-management/cart/custom-order-reference-cart-page.ts
@@ -9,27 +9,12 @@ import { CustomOrderReferenceCartRepository } from './custom-order-reference-car
 export class CustomOrderReferenceCartPage extends YvesPage {
   @inject(REPOSITORIES.CustomOrderReferenceCartRepository) private repository: CustomOrderReferenceCartRepository;
 
-  protected PAGE_URL = '/cart';
-
-  visitCartWithItems = (): void => {
-    this.visit();
-    this.repository.getCartUpsellingAjaxLoader().should('be.visible');
-    this.repository.getCartUpsellingAjaxLoader().should('be.not.visible');
-  };
-
   addCustomOrderReferenceInput = (reference: string): void => {
     this.repository.getCustomOrderReferenceInput().type(reference);
     this.repository.getCustomOrderReferenceSubmitButton().click();
-
-    this.repository.getCustomOrderReferenceAjaxLoader().should('be.visible');
-    this.repository.getCustomOrderReferenceAjaxLoader().should('be.not.visible');
   };
 
   getCustomOrderReferenceInput = (): Cypress.Chainable => {
     return this.repository.getCustomOrderReferenceInput();
-  };
-
-  getCustomOrderReferenceSubmitButton = (): Cypress.Chainable => {
-    return this.repository.getCustomOrderReferenceSubmitButton();
   };
 }
