+++ cypress/support/pages/yves/order-management/cart/custom-order-reference-cart-repository.ts
@@ -1,6 +1,4 @@
 export interface CustomOrderReferenceCartRepository {
-  getCartUpsellingAjaxLoader(): Cypress.Chainable;
   getCustomOrderReferenceInput(): Cypress.Chainable;
   getCustomOrderReferenceSubmitButton(): Cypress.Chainable;
-  getCustomOrderReferenceAjaxLoader(): Cypress.Chainable;
 }
