+++ cypress/support/pages/yves/order-management/cart/repositories/suite-custom-order-reference-cart-repository.ts
@@ -3,12 +3,8 @@ import { CustomOrderReferenceCartRepository } from '../custom-order-reference-ca
 
 @injectable()
 export class SuiteCustomOrderReferenceCartRepository implements CustomOrderReferenceCartRepository {
-  getCartUpsellingAjaxLoader = (): Cypress.Chainable =>
-    cy.get('[data-qa="component cart-upselling"] [data-qa="component ajax-loader"]', { timeout: 10000 });
   getCustomOrderReferenceInput = (): Cypress.Chainable =>
     cy.get('[data-qa="component order-custom-reference-form"] input[type=text]');
   getCustomOrderReferenceSubmitButton = (): Cypress.Chainable =>
     cy.get('[data-qa="component order-custom-reference-form"] button[type=submit]');
-  getCustomOrderReferenceAjaxLoader = (): Cypress.Chainable =>
-    cy.get('[data-qa="component ajax-loader order-custom-reference-ajax-loader"]');
 }
